+++ src/main/java/nextstep/auth/application/OAuth2Client.java
@@ -1,0 +1,8 @@
+package nextstep.auth.application;
+
+import nextstep.auth.application.dto.OAuth2ProfileResponse;
+
+public interface OAuth2Client {
+    String requestToken(String code);
+    OAuth2ProfileResponse requestProfile(String accessToken);
+}
