+++ src/main/java/nextstep/auth/application/dto/OAuth2LoginRequest.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto;
+package nextstep.auth.application.dto;
 
 public class OAuth2LoginRequest {
     private String code;
