+++ src/main/java/nextstep/auth/application/dto/OAuth2ProfileResponse.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto;
+package nextstep.auth.application.dto;
 
 public class OAuth2ProfileResponse {
     private String email;
