--- src/main/java/nextstep/member/application/OAuth2Client.java
@@ -1,8 +1,0 @@
-package nextstep.member.application;
-
-import nextstep.member.application.dto.OAuth2ProfileResponse;
-
-public interface OAuth2Client {
-    String requestGithubToken(String code);
-    OAuth2ProfileResponse requestGithubProfile(String accessToken);
-}
