+++ src/test/java/nextstep/member/acceptance/AuthAcceptanceTest.java
@@ -3,9 +3,9 @@ package nextstep.member.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.member.application.dto.GithubAccessTokenResponse;
-import nextstep.member.application.dto.OAuth2ProfileResponse;
-import nextstep.member.application.dto.TokenResponse;
+import nextstep.auth.application.dto.GithubAccessTokenResponse;
+import nextstep.auth.application.dto.OAuth2ProfileResponse;
+import nextstep.auth.application.dto.TokenResponse;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import nextstep.utils.AcceptanceTest;
