+++ src/test/java/nextstep/member/acceptance/AuthSteps.java
@@ -3,7 +3,7 @@ package nextstep.member.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.member.application.dto.OAuth2LoginRequest;
+import nextstep.auth.application.dto.OAuth2LoginRequest;
 import org.springframework.http.MediaType;
 
 public class AuthSteps {
