+++ src/test/java/nextstep/member/application/GithubClientTest.java
@@ -1,6 +1,7 @@
 package nextstep.member.application;
 
-import nextstep.member.application.dto.OAuth2ProfileResponse;
+import nextstep.auth.application.GithubClient;
+import nextstep.auth.application.dto.OAuth2ProfileResponse;
 import nextstep.utils.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,7 +19,7 @@ public class GithubClientTest extends AcceptanceTest {
     @DisplayName("깃허브 accessToken 요청")
     @Test
     void requestGithubToken() {
-        String accessToken = githubClient.requestGithubToken(사용자1.getCode());
+        String accessToken = githubClient.requestToken(사용자1.getCode());
 
        assertThat(accessToken).isEqualTo(사용자1.getAccessToken());
     }
@@ -26,7 +27,7 @@ public class GithubClientTest extends AcceptanceTest {
     @DisplayName("깃허브 profile 요청")
     @Test
     void requestGithubProfile() {
-        OAuth2ProfileResponse res = githubClient.requestGithubProfile(사용자1.getAccessToken());
+        OAuth2ProfileResponse res = githubClient.requestProfile(사용자1.getAccessToken());
 
         assertThat(res.getEmail()).isEqualTo(사용자1.getEmail());
         assertThat(res.getAge()).isEqualTo(사용자1.getAge());
