+++ src/test/java/nextstep/member/application/MemberServiceTest.java
@@ -2,7 +2,7 @@ package nextstep.member.application;
 
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
-import nextstep.member.application.dto.OAuth2ProfileResponse;
+import nextstep.auth.application.dto.OAuth2ProfileResponse;
 import nextstep.member.domain.Member;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
